# Conflicts:
#	src/Erp/NotificationBundle/Resources/views/UserNotification/create.html.twig

+++ src/Erp/NotificationBundle/Form/Type/NotificationType.php
@@ -31,7 +31,7 @@ class NotificationType extends AbstractType
                     ],
                     'constraints' => [
                         new Assert\NotBlank(),
-                        new Assert\Type('integer'),
+                        new Assert\Type('numeric'),
                         new Assert\GreaterThanOrEqual(['value' => 1]),
                     ],
                 ]
